+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fixed a stack overflow occurring when parsing CLI options that occurred on
   debug Windows builds of `sprocket` (https://github.com/stjude-rust-labs/sprocket/pull/1224).
+* Nested inputs are now shell-expanded in `run` ([#1231](https://github.com/stjude-rust-labs/sprocket/pull/1231)).
 
 ## 0.31.0 - 2026-09-16